pyproject.toml
[tool.poetry]
name = "im2latex"
version = "0.1.0"
description = ""
authors = ["Matthijs <matthijs.vanderlende@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.47.0"
torch = "^2.5.1"
numpy = "^2.2.0"
evaluate = "^0.4.3"
datasets = "^3.2.0"
pillow = "^11.0.0"
tqdm = "^4.67.1"
peft = "^0.14.0"
loguru = "^0.7.3"
hydra-core = "^1.3.2"
nltk = "^3.9.1"
python-dotenv = "^1.0.1"
wandb = "^0.19.1"
pyperclip = "^1.9.0"
pyqt5 = "^5.15.11"
streamlit = "^1.41.1"
mss = "^10.0.0"
pyautogui = "^0.9.54"
grad-cam = "^1.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
